--- a/.travis.yml
+++ b/.travis.yml
@@ -1,13 +1,12 @@
 language: go
 
 go:
-  - 1.5.3
+  - 1.7
 
 install:
   - go get -u github.com/golang/lint/golint
   - go get -u github.com/kisielk/errcheck
   - go get -u golang.org/x/tools/cmd/cover
-  - go get -u golang.org/x/tools/cmd/vet
   - test -z "$(go fmt)"
   - test -z "$(golint)"