This change reduces allocations by about 50%. The next step is optimizing the call to `h.Sum`. Currently we always pass `nil`, so a new slice is allocated each time. Ideally we could reuse memory from elsewhere in the tree. This is trivial if we only care about calculating the root (just reuse child memory when calculating the parent hash), but non-trivial when proofs are required. This suggests the possibility of a separate `Root` codepath that is more aggressively optimized. Of course, it would be better if we could apply this optimization when performing proofs as well, but I don't understand `Prove` well enough to attempt it myself.