@lukechampine 2017-01-19
Old benchmark:
```
BenchmarkTree64_4MB-4         20	  85377604 ns/op	25297283 B/op	  589829 allocs/op
BenchmarkTree4k_4MB-4        100	  12830937 ns/op	 5032461 B/op	    9226 allocs/op
```
New benchmark:
```
BenchmarkTree64_4MB-4         20	  74705171 ns/op	10617048 B/op	  393219 allocs/op
BenchmarkTree4k_4MB-4        100	  11461825 ns/op	  166080 B/op	    6150 allocs/op
```