@DavidVorick 2017-01-19
Test is no longer passing for special-case inputs. The result of these inputs is strictly defined in RFC 6962, and furthermore I think would be a hardfork to change.

Need to make sure you are returning correct values in all cases.