@lukechampine 2017-01-19
The RFC says "The hash of an empty list is the hash of an empty string," but `Root` doesn't appear to conform to that (it returns `nil` instead). I suppose we need to preserve that behavior.