@DavidVorick 2017-01-19
oh dear. I've always read that as "is the empty string". Crap.