@lukechampine 2017-01-19
`sum` never returns `nil` in practice because it is never supplied a `nil` argument; the input bytes are always prefaced by a `0` or `1`. So the only time `sum` could return `nil` is during testing.