@DavidVorick 2017-01-19
ok. Going to hold off on merging until v1.1.0 has been released.